Update tests related to ListTemplateFieldDefinitionsByTemplateId
